requirements.txt
pyside6>=6.5
psycopg2-binary>=2.9
sqlparse>=0.4
